// src/app/main/pages/tax-income/upload/models/content.interface.ts
export interface ITaxIncomeUploadDetail {
    makerDate?: any;
    authStatus?: any;
    tenantCode?: any;
    exceptionContentExt?: any;
    content: {
        buEmail?: any;
        pattern?: any;
        buAddress?: any;
        empTaxincomeWithheld?: any;
        empPersonaltaxWithheld?: any;
        empName?: any;
        empTaxincomeCalculation?: any;
        makerId?: any;
        buId?: any;
        companyCode?: any;
        empNationality?: any;
        empPayFromMonth?: any;
        buCode?: any;
        buName?: any;
        empPayYear?: any;
        buTaxcode?: any;
        empCode?: any;
        serial?: any;
        empIsResident?: any;
        buPhone?: any;
        rowIndex?: any;
        empPayToMonth?: any;
        empTimePay?: any;
        empTaxcode?: any;
        empContactAddress: any;
        empContactPhone: any;
        empIncomeType: any;
    };
    modNo?: any;
    recordStatus?: any;
    checkerDate?: any;
    exceptionContent: {
        empTaxcode?: any;
    };
    checkerId?: any;
    id?: any;
    ftsValue?: any;
    makerId?: any;
    status?: any;
    smtbProcessUpload?: any;
    //
    checked?: any;
    showCheckBox?: any;
}